[profile] move traversal PFFUN into its own phase
Previously the traversal PFFUN, which annotates function applications
for the profiler, was toward the end of the optimisation phase. This is
fine except that the annotations, stored as N_annotate as a statement of
an N_assign, were being moved around by later traversals (CUDA and Mem),
meaning that some counted operations were being tracked by the wrong
function record. We now move this traversal to be at the end of the Mem
phase, into is own phase which we call Iprof (insert profiling
annotations).