[emr-cuda-lift] change which candidate we choose
When lifting h2d out of loop (which as introduced as part of EMR loop
optimisation), we need to find a suitable device type array to be placed
in the arguments of the loop recursive call. Previously we assumed we
could use the LHS of the h2d, but this was erronous. Instead we now scan
though our loop function for all WL LHS, and the RHS of d2h, and use one
of these instead. We make sure not not select candidate that is a RC/ERC
or an argument to function application.