[cuda-cuade] fix rare situration of paired h2d/d2h
It can happen that we reference both the host and device memory of a
given array though some function application. In such a situation we
would could have both a d2h and h2d which assign to these arrays. If we
are not careful, the CUADE optimisation could be overly eager to move
both the h2d statement up and the d2h statement down, leading in the
worst case to the statements being swapped. This leads to
'uninitialised' related C-compiler errors and segfaulting programs.