[cuda-cuade] add missing handling of lifted ERCs
We did not handle the case where we have a lifted ERC value as the RHS
of a h2d. This means that we would capture the N_assign, but then _drop
it_ as we never found it's RHS being assigned to (or being an argument
to the current N_fundef). This commit fixes this by search through the
vardecs...