[cuda-cuade] better handling of N_cond
previously we didn't do anything with the LUTs when traversing the N_cond,
no we handle it similarly to the N_with - we carry with us its preceding
N_assign, and use this whenever we find a N_id with a d2h_end as its
assignment. Because we are using LUTs, once we traverse out of the
N_cond and find another reference to an already seen N_id, we just
update it.