[mt] elide SAC_MT_DEFINE
This macro was used to store some global declarations for the MT
backend. This was made redundant some time ago, when shifting the
scheduler variables into the beehive. SAC_MT_DEFINE was not removed
from the code generation or the schedule.h, leading to a duplication
of declarations across compilation units.

This issue was not noticed until GCC 10, where `-fno-common` become
default, disallowing duplicate global variable definitions.