skipping potential recursive calls in -profile f elided
This triggered a test to fail on centos only :-)
The problem is: if there is only one function and we skip the first one, we are out of bounds! Ergo: anything can happen!
There was a comment that this would be needed for function 0 (main) and function 0 only, but as far as I understand,
this is not the case. In a trivial example (test-scalar-sum-profile) even main has a registered call, namely its initial launch.