fixed the offending potential reuse of the neutral element in non-scalar fold-WLs
By setting the rc in the stack copy of the descriptor to 2 we are safe now. I also injected tracing info so that it easier to see what is going on just from the trace.
As a consequence of this, the MT version now leaks one copy of the neutral element! This needs another fix.
Finally, I added some comments in MTRMI to explain what exactly it does and to understand that traversal more quickly :-)