added full support for nested types in the typechecker
actually not much was missing here. However, the treatment of T_hidden received a massive conceptual overhaul.
This was triggered by the observation that SACarg needs to be treated like a nested data structure.....
When ironing that out in compile.c we can equally well make sure we add proper support for nesting throughout....