change gitlab-ci to use docker images directly
previously we set the docker image in the runner, meaning that we
could not control what version of the images were being used. This could
at time lead to things breaking.

We now have to set the version explicitly in the config file for the CI,
which should allow use to *test* newer images before deploying them.