[cmake] mark flag as CLANG only
In GCC, any `Wno-...` flag is silently ignored if it is invalid *UNLESS*
there is some other diagnostic output. The effect is that we cannot
strictly test for any `Wno-...` flags without seriously overhauling how
we do the flag check test (for diagnostic output?). The flag in question
anyway does not exist in GCC, but in CLANG, so it has been deactivated
except in the latter case.