added debug option -dgpukernel and brushed phase reporting in cuda phase
Now, there is a global variable global.gpukernel which is per default false but can be set to true in the debug version using -dgpukernel
It is meant to provide extensive checking of the gpukernel pragma translation. As this is adding significant overhead, it is only
available as debug option and not as part of -check g

In the cuda phase, we now indicate if some phases are (hybrid only) or (non-hybrid). All non-annotated subphases in cuda are
executed in both backends: cuda and cuda-hybrid.