ignore ld warnings about 'built for newer version'
At some point Apple started supporting multiple XCode installs. By
default the latest version is used, unless the user specifies otherwise
using the `MACOSX_DEPLOYMENT_TARGET` environment variable and the
`-mmacosx-version-min=` clang flag. At linking time, LD does some check
to see if the library was compiled with the same deployment target, if
not it gives a warning.

These warnings are somewhat nonsensical as far as I have been able to
understand (no offical documentation exists for this), and so it seems
safe to just ignore this. For the sake of the CI this is what we do.