[build] create .sac2crc directory
Previously we would error out if it did not exists, expecting the
user to create it for us. This pattern was not good, and so now we've
changed this to instead check if the path points to a file, if so then
we error out (as it might be the case that the user is still using the
old sac2crc setup). Otherwise we just continue, and copy over the new
sac2crc prelude file.