[cmake] switch to CUDAToolkit find module
the original cmake CUDA find module was depreciated with version 3.10.
The new module is a bit more slick, and seems to be better at find the
correct paths to CUDA toolkit.

Additionally, with archlinux, some applications (like CUDA) are
installed under /opt or some other non-standard path. This means that
any binaries (and even shared libraries) can be missing from PATH (and
other environment variables). This changes forces the /etc/profile to be
loaded, which resolves this issue. It is normally loaded when using a
'login' shell, but gitlab-runners don't do this.