[ctest] add REQFEATURES keyword for feature depend
For all test-*.sac unit tests we now can add the `REQFEATURES|` keyword
with a list of ENABLE_* flags to mark a test depending on those
features. In this way, we can deactivate tests from running.

This is different from the ISALLOPTIONSON script, which does not
actually remove the test from the CTest run list. Instead it is
used as part of a conditional to change the test run parameters.