[proj] move to ISC license; make sac2c OSS
With this commit we move the sac2c project to a permissive
OSS license called the Internet Software Consortium (ISC) license,
which is similar to the simpler MIT/BSD licenses, but with a more
concise text.

Additionally we create a 'CONTRIBUTORS' file which includes all the
names of individuals who contributed to the source code of the project.

Finally, at compile time, the license and contributors lists is baked
into the compiler directly replacing the fixed string messages in
`usage.c`.