[macos] move to using CC binary; drop fixed sysroot
Using CMAKE to set the sysroot statically at the time we build
sac2c sets limits on the packaged version of sac2c. CMAKE sysroot
is versioned to match the XCode/SDK used to compile sac2c. This
versioned path **only** exists for a specific version of XCode. On other
setups, newer/older version of MacOS, and even newer/older version of
XCode, this path is probably invalid.

The solution is to use /usr/bin/cc instead, which is a wrapper around
clang that correctly sets the sysroot path information (and other
details besides).

Additionally this commit:
* move to cmake 3.19 base support (for better support on Apple M1)
* disable (manually) FAT binary generation (for Apple systems)
* use -mmacosx-version-min flag to correct set minimum support SDK
  version
* extend str.c; correct module/function comments; add STRstrip function
  to remove newline/space characters
* extend unit tests for strings (STRstrip)
* add platform/macos.c util functions (NEEDED?)
* add scripts/get-xcode-version.sh script (NEEDED?)