Potential fix for ci breaking with new warnings
In the old system, warning messages were obtained in the format of `warning: <first line of warning_message>`.

In the new system, warning messages are displayed on lines that follow lines containing nothing but "Warning:".  
The first grep setup grabs the warnings and the line after each warning. The first lines containing "Warning:" are then removed by the second grep. The third grep now greps on the first line of the warning message prefixed by two spaces, which should be close enough to the original behaviour.