added missing renaming in MOWL handling
added the renaming, extended the explanation and streamlined the code modification.
also changed the default verbosity for the debug version to v2...
and added a few tests

one of the tests revealed that we, in some cases, do get self-assignment warnings from prop-obj -in/ out code generation.
I suppressed them on MAC for now. If they pop up on gcc based systems as well, we shall do the same there.