[emrl] filter out used ERCs from pool
Ahhh, for whatever reason the case that there might be several
identically shaped ERCs needed for a loop was not correctly handled
here. This led to the **same** candidate being used for all (of the
same shape) WLs in a loop function, aaaaahahhhh. With this commit we
filter out the candidate from the pool of possible candidates.