fix missing minimum deployment version on MacOS
MacOS introduced the notion of minimum deployment target/version
which sets out a minimum system version (MacOS X version, e.g. 10.11)
and similar that _guarantees_ that the binary will work on that system.

CMAKE supports setting the correct flags for this **but** seems to fail
to determine the minimum deployment version. Its not clear to me why this
is just that I have seen this happen consistently on systems with
multiple SDKs versions installed.

In such a case we manually set it to version 10.14 (Mojave) now.