fix cmake CMP0114 policy, set to NEW
The primary change in behaviour is with how a *_Step dependency is
handled. Previously, all *_Step, would automatically depend on all
targets given in `DEPENDS`, even if this didn't make much sense. Issues
with race-conditions and repeatedly calling *_Step could happen due to
this. The new behaviour sets out that each *_Step is independent
**unless** an explicit `DEPENDS` is set.