disable use-after-free warning for memalign()
the warning is likely a false positive, the underlying pointer structure
is a three part union with padding. The static analysis likely can't
fully expand this (and really I don't see how it could), and so can't
accurately track the effect of the pointer offsets. We disable the
warning for this line...